+++ CarrierWave/Models/ActivityType.swift
@@ -15,6 +15,7 @@ enum ActivityType: String, Codable, CaseIterable {
     case dailyStreak
     case potaDailyStreak
     case personalBest
+    case workedFriend
 
     // MARK: Internal
 
@@ -31,6 +32,7 @@ enum ActivityType: String, Codable, CaseIterable {
         case .dailyStreak: "flame.fill"
         case .potaDailyStreak: "flame.fill"
         case .personalBest: "chart.line.uptrend.xyaxis"
+        case .workedFriend: "person.line.dotted.person.fill"
         }
     }
 
@@ -47,6 +49,7 @@ enum ActivityType: String, Codable, CaseIterable {
         case .dailyStreak: "Daily Streak"
         case .potaDailyStreak: "POTA Streak"
         case .personalBest: "Personal Best"
+        case .workedFriend: "Worked Friend"
         }
     }
 }